Move airtouch4 coordinator to its own file
